Terminating ';' optional.
Prior to this commit there was a requirement that a `;` was used
to denote the end of a production. This had the potential to
lead to incompatibilities with grammars that weren't doctored for
use with this library. After this commit, the `;` remains an optional
end of production indicator but is no longer required.

End of production is inferred by the end of an expression.
End of an expressions is denoted by the following:
- eol
- `;`
- `|`
- the start of the next production: `<...> ::=`