Add logic to handle incomplete input.
Some of the logic implemented in order to validate input and
provide appropriate errors resulted in reverting some of the
parser revisions captured in an earlier commit. Lookaheads
are leaned on in order to verify the end of an expression.

If the parser sees `|` that isn't followed by another expression
we can provide an error.

If the parser sees the lhs pattern for a production it can recognize
it rather than attempting to parse it as simply another nonterminal.

Also, the tests that fail on what we're calling "incomplete" input
correspond to a nom error rather than a nom incomplete. If our
parser does its job right the user should never see an incomplete
which would indicate we consumed something we expect later rather than
just the absense of an expected specific token.