Revise node impls to remove from_parse
Prior to this commit from_str was an alias for another
trait called from_parse. This commit removes from_parse
so there is only from_str. It exists as a custom trait
of each node so the user isn't required to
`use std::str::FromStr` but that trait is implemented
for each node as well so a user may use these types in
generic functions that require that trait.