Trim BNF for BNF grammar and allow stack growth
Prior to this commit the BNF for BNF had a rule that catered to
deep recursion without, to me, adding any real value as sensical syntax
despite that it is how the wikipedia source defines the grammar.

Although the revision to the gramamr makes our failure case much less
likely, I also added some logic to allow the stack to grow for deep
recursion in order to give grammars that might require it more room to
generate. The fact that ours had failed with regularity made me think
there is a use case for allowing more stack within a threshold. That
threshold is exercised by our `infinite_loop` test.