Added error for hitting recursion limit
This commit contains a few revisions.

First, a new error kind was introduced to indicate that
the generation error isn't generic but the result of the
limits of recursion being reached. This allows the
property tests for the Grammar type to disregard problematic
sentence generation, it is testing that what it gets is parsable.

In order to manage meeting recursion limits being met in the
property tests, a new use case was accounted for, one that I
belive implementaion for is useful. Logic was added to allow the
"from_str" trait for bnf types to be parallel to Type::new()
in all cases except the Term type. Because the Term type does
not have a sensical way to implement new(), "from_str" for Term
types corresponds to a Terminal rule for the empty string
which is allowed in parsable gramamrs.

As this commit represents that addition of a new error kind
it seemed that there is enough variety now to make our Error
type public in order for users can match on them.