First commit to support ISO 14977 compliant impl of EBNF
Using pest because I find nom to not be robust enough to express.
I might be wrong and it turns out to be easier to implement with nom.
But there are some things that are simply not easily expressable
using nom parses. For example, how to express nested group-sequence?

For now I am quite tired, this seems to be a lot more complicated that
I thought it would be. Turns out this EBNF is quite powerful. I originally
intended to use this as a verification tool for DNS names but I think I will just
use pest after learning how comprehensive it is. Or just manual parsing.

I will leave this to someone more capable.

Signed-off-by: Hanif Bin Ariffin <hanif.ariffin.4326@gmail.com>
Signed-off-by: Hanif Ariffin <hanif.ariffin.4326@gmail.com>