Earley predict on complete input
An original (maybe overeager!) optimization was included with Earley
parsing. If there was no remaining input text, then no more predictions
from the grammar were attempted. In most cases, this saves Earley a
some work. But the BNF crate supports empty production
rules. This invalidates this optimization, because an empty production
may still be successful even with no remaining input text.

This commit only removes this optimization built on that false
assumption. It would be possible to reintroduce this improvement, but
*only* for grammars without any empty productions.